move header components to separate file
